+++ renderer/units/Unicode Data Finder/import.js
@@ -56,7 +56,7 @@ let defaultFolderPath;
 //
 module.exports.start = function (context)
 {
-    const { remote } = require ('electron');
+    const { clipboard, Menu, shell } = require ('electron').remote;
     //
     const path = require ('path');
     //
@@ -272,13 +272,13 @@ module.exports.start = function (context)
             }
             else
             {
-                remote.shell.beep ();
+                shell.beep ();
             }
         }
     );
     //
     let nameResultsMenu =
-    remote.Menu.buildFromTemplate
+    Menu.buildFromTemplate
     (
         [
             {
@@ -287,7 +287,7 @@ module.exports.start = function (context)
                 {
                     if (currentCharactersByName.length > 0)
                     {
-                        remote.clipboard.writeText (currentCharactersByName.join (""));
+                        clipboard.writeText (currentCharactersByName.join (""));
                     }
                 }
             },
@@ -445,13 +445,13 @@ module.exports.start = function (context)
             }
             else
             {
-                remote.shell.beep ();
+                shell.beep ();
             }
         }
     );
     //
     let matchResultsMenu =
-    remote.Menu.buildFromTemplate
+    Menu.buildFromTemplate
     (
         [
             {
@@ -460,7 +460,7 @@ module.exports.start = function (context)
                 {
                     if (currentCharactersByMatch.length > 0)
                     {
-                        remote.clipboard.writeText (currentCharactersByMatch.join (""));
+                        clipboard.writeText (currentCharactersByMatch.join (""));
                     }
                 }
             },
@@ -729,7 +729,7 @@ module.exports.start = function (context)
                 }
                 else
                 {
-                    remote.shell.beep ();
+                    shell.beep ();
                 }
             }
             else
@@ -773,7 +773,7 @@ module.exports.start = function (context)
     );
     //
     let blockResultsMenu =
-    remote.Menu.buildFromTemplate
+    Menu.buildFromTemplate
     (
         [
             {
@@ -782,7 +782,7 @@ module.exports.start = function (context)
                 {
                     if (currentCharactersByBlock.length > 0)
                     {
-                        remote.clipboard.writeText (currentCharactersByBlock.join (""));
+                        clipboard.writeText (currentCharactersByBlock.join (""));
                     }
                 }
             },